pyproject.toml
[project]
name = "document-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "fastapi>=0.121.0",
    "filetype>=1.2.0",
    "google-cloud>=0.34.0",
    "google-cloud-vision>=3.11.0",
    "google-genai>=1.48.0",
    "openai>=2.7.1",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "tesseract>=0.1.3",
    "uvicorn>=0.38.0",
]
